+++ convnetGoLib/tap_ctrl.go
@@ -7,7 +7,6 @@ import (
 	"log"
 	"net"
 	"os/exec"
-	"strings"
 
 	"github.com/songgao/water"
 	"github.com/songgao/water/waterutil"
@@ -27,12 +26,7 @@ func TapInit() {
 		client.Mymac = GetMymac(ifce.Name())
 
 		client.g_ifce = ifce
-		defer func() {
-			if err := ifce.Close(); err != nil {
-				client.g_ifce = nil
-				Log(err)
-			}
-		}()
+		defer teardownIfce(ifce)
 
 	}
 
@@ -97,18 +91,6 @@ func startRead(ifce *water.Interface) (dataChan <-chan []byte, errChan <-chan er
 
 func startPing(dst net.IP, _ bool) {
 	if err := exec.Command("ping", "-n", "4", dst.String()).Start(); err != nil {
-		Log(err)
-	}
-}
-
-func setupIfce(ipNet net.IPNet, dev string) {
-	sargs := fmt.Sprintf("interface ip set address name=REPLACE_ME source=static addr=REPLACE_ME mask=REPLACE_ME gateway=none")
-	args := strings.Split(sargs, " ")
-	args[4] = fmt.Sprintf("name=%s", dev)
-	args[6] = fmt.Sprintf("addr=%s", ipNet.IP)
-	args[7] = fmt.Sprintf("mask=%d.%d.%d.%d", ipNet.Mask[0], ipNet.Mask[1], ipNet.Mask[2], ipNet.Mask[3])
-	cmd := exec.Command("netsh", args...)
-	if err := cmd.Run(); err != nil {
 		Log(err)
 	}
 }
